requirements.txt
streamlit
pdfminer.six
numpy
scikit-learn
sentence-transformers
pandas
matplotlib
